Gives the inconsistencies between the weights of small and big sample fish and the sum of the small and big weights in the associated well

The purpose of the check_distribution_inspector function is to provide a table of data that contains an inconsistency between the small and large sample weights and the sum of the small and big weights of the associated well

check_distribution_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  dataframe4,
  output,
  species = c("SKJ"),
  weight_category_small = c("W-1"),
  weight_category_big = c("W-2"),
  weight_category_unknown = c("W-9")
)

Arguments

dataframe1

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_distribution_inspector () function.

dataframe2

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_distribution_inspector () function.

dataframe3

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_distribution_inspector () function.

dataframe4

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_distribution_inspector () function.

output

character expected. Kind of expected output. You can choose between "message", "report" or "logical".

species

character expected. Default values: c("SKJ"). Vector of species categorized as small if weight category information is missing

weight_category_small

character expected. Default values: c("W-1"). Vector of small weight category codes

weight_category_big

character expected. Default values: c("W-2"). Vector of big weight category codes

weight_category_unknown

character expected. Default values: c("W-9"). Vector of unknown weight category codes

Value

The function returns a character with output is "message", a data.frame with output is "report", a logical with output is "logical"

Details

The input dataframe must contain all these columns for the function to work :

sample_id

sample_well

trip_id

sample_smallsweight

sample_bigsweight

well_id

well_label

trip_id

wellactivity_id

well_id

wellactivityspecies_id

wellactivity_id

weightcategory_code

species_fao_code

wellactivityspecies_weight

Examples

dataframe1 <- data.frame(sample_id = c("1", "2", "3"),
                         sample_well = c("well_1", "well_2", "well_3"),
                         trip_id = c("1", "1", "1"),
                         sample_smallsweight = c(6, 25, 14),
                         sample_bigsweight = c(12, 0, 9))
dataframe2 <- data.frame(well_id = c("1", "2", "3"),
                         well_label = c("well_1", "well_2", "well_3"),
                         trip_id = c("1", "1", "1"))
dataframe3 <- data.frame(wellactivity_id = c("1", "2", "3", "4"),
                         well_id = c("1", "1", "2", "3"))
dataframe4 <- data.frame(wellactivityspecies_id = c("1", "2", "3", "4", "5", "6", "7"),
                         wellactivity_id = c("1", "1", "1", "2", "3", "4", "4"),
                         weightcategory_code = c("W-1", "W-9", "W-2", "W-2", "W-1", "W-2", "W-9"),
                         species_fao_code = c("BET", "SKJ", "SKJ", "ALB", "SKJ", "BET", "BET"),
                         wellactivityspecies_weight = c(4, 2, 7, 5, 25, 9, 14))
check_distribution_inspector(dataframe1, dataframe2, dataframe3, dataframe4, output = "report")
#> # A tibble: 3 × 7
#> # Groups:   sample_id [3]
#>   sample_id logical sample_smallsweight sample_bigsweight sample_well
#>   <chr>     <lgl>                 <dbl>             <dbl> <chr>      
#> 1 1         TRUE                      6                12 well_1     
#> 2 2         TRUE                     25                 0 well_2     
#> 3 3         FALSE                    14                 9 well_3     
#> # ℹ 2 more variables: weight_small_total <dbl>, weight_big <dbl>