Gives the inconsistencies between the fishing area declared and calculated for the activity

The purpose of the check_eez_inspector function is to provide a table of data that contains an inconsistency between the fishing area declared and calculated with position for the activity fishing

check_eez_inspector(
  dataframe1,
  dataframe2,
  output,
  activity_crs = 4326,
  international_waters_code = "XIN",
  vessel_activity = c("6")
)

Arguments

dataframe1

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_weight_inspector () function.

dataframe2

data.frame expected. Layer to containing the eez shapefile (example cf : Flanders Marine Institute (2023). Maritime Boundaries Geodatabase: Maritime Boundaries and Exclusive Economic Zones (200NM), version 12. Available online at https://www.marineregions.org/. https://doi.org/10.14284/632)

output

character expected. Kind of expected output. You can choose between "message", "report" or "logical".

activity_crs

numeric expected. Default values: 4326. Coordinate Reference Systems for the position activity

international_waters_code

character expected. Default values: "XIN". iso3 code corresponding to international waters

vessel_activity

character expected. Default values: c("6"). Vector of inventory of codes for activities that must have a zee zone

Value

The function returns a character with output is "message", two data.frame with output is "report" (the first without geographical location and the second with geographical location), a logical with output is "logical"

Details

The input dataframe must contain all these columns for the function to work :

activity_id

vesselactivity_code

fpazone_code

fpazone_country_iso3

activity_position

ISO_TER1

ISO_TER2

ISO_TER3

geometry

Examples

dataframe1 <- data.frame(activity_id = c("1", "2", "3", "4"),
              vesselactivity_code = c("6", "6", "6", "6"),
              fpazone_code = c("SYC", "XSG", "XIN", "SYC"),
              fpazone_country_iso3 = c("SYC", "XXX", "XIN", "SYC"),
              activity_position = c("POINT (1 1)", "POINT (4 3)", "POINT (-1 -1)", "POINT (-1 -1)"))
dataframe2 <- sf::st_sf(data.frame(ISO_TER1 = c("SYC", "XSG"),
          ISO_TER2 = c(NA, NA),
          ISO_TER3 = c(NA, NA),
          geometry = sf::st_sfc(sf::st_polygon(list(rbind(c(0,0), c(2,0), c(2,2), c(0,2), c(0,0)))),
                                sf::st_polygon(list(rbind(c(3,3), c(3,5), c(5,5), c(5,3), c(3,3)))),
                                crs = 4326)))
check_eez_inspector(dataframe1, dataframe2, output = "report")
#> [[1]]
#>   activity_id vesselactivity_code fpazone_code fpazone_country_iso3 logical
#> 1           1                   6          SYC                  SYC    TRUE
#> 2           2                   6          XSG                  XXX    TRUE
#> 3           3                   6          XIN                  XIN    TRUE
#> 4           4                   6          SYC                  SYC   FALSE
#> 
#> [[2]]
#>   activity_id vesselactivity_code fpazone_code fpazone_country_iso3
#> 1           1                   6          SYC                  SYC
#> 2           2                   6          XSG                  XXX
#> 3           3                   6          XIN                  XIN
#> 4           4                   6          SYC                  SYC
#>   activity_position logical activity_crs
#> 1       POINT (1 1)    TRUE         4326
#> 2       POINT (4 3)    TRUE         4326
#> 3     POINT (-1 -1)    TRUE         4326
#> 4     POINT (-1 -1)   FALSE         4326
#>