Gives the inconsistencies between the fishing times or sea times indicated for the route and activities carried out

The purpose of the check_time_route_inspector function is to provide a table of data that contains an inconsistency between the fishing times or sea times indicated for the route and activities carried out

check_time_route_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  output,
  max_seatime = 24,
  max_fishingtime = 13,
  vesselactivity_seatime = c("1", "2", "3", "4", "6", "6", "8", "9", "10", "11", "12",
    "13", "13", "13", "13", "13", "13", "14", "15", "17", "18", "19", "20", "22", "23",
    "24", "25", "26", "27", "29", "30", "31", "32", "36", "37", "38", "39", "50", "99",
    "101", "101", "102", "102", "103"),
  objectoperation_seatime = c(NA, NA, NA, NA, "99", NA, NA, NA, NA, NA, NA, "1", "2",
    "4", "6", "8", "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
    NA, NA, NA, NA, NA, NA, "99", NA, "99", NA, NA),
  vesselactivity_fishingtime = c("2", "3", "4", "6", "6", "8", "12", "13", "13", "13",
    "13", "13", "13", "14", "15", "17", "19", "20", "23", "25", "26", "27", "29", "30",
    "31", "102", "102"),
  objectoperation_fishingtime = c(NA, NA, NA, "99", NA, NA, NA, "1", "2", "4", "6", "8",
    "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "99", NA)
)

Arguments

dataframe1

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_time_route_inspector () function.

dataframe2

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_time_route_inspector () function.

dataframe3

data.frame expected. Csv or output of the function data_extraction, which must be done before using the check_time_route_inspector () function.

output

character expected.Kind of expected output. You can choose between "message", "report" or "logical".

max_seatime

numeric expected. Default values: 24. Maximum valid time for sea time

max_fishingtime

numeric expected. Default values: 13. Maximum valid time for fishing time

vesselactivity_seatime

character expected. Default values: c("1", "2", "3", "4", "6", "6", "8", "9", "10", "11", "12", "13", "13", "13", "13", "13", "13", "14", "15", "17", "18", "19", "20", "22", "23", "24", "25", "26", "27", "29", "30", "31", "32", "36", "37", "38", "39", "50", "99", "101", "101", "102", "102", "103"). Code activities. First criterion for identifying activities that must have sea time

objectoperation_seatime

character expected. Default values: c(NA, NA, NA, NA, "99", NA, NA, NA, NA, NA, NA, "1", "2", "4", "6", "8", "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "99", NA, "99", NA, NA). Code object operation. Second criterion for identifying activities that must have sea time (Indicate NA if you want the code object operation to be missing)

vesselactivity_fishingtime

character expected. Default values: c("2", "3", "4", "6", '6', "8", "12", "13", "13", "13", "13", "13", "13", "14", "15", "17", "19", "20", "23", "25", "26", "27", "29", "30", "31", "102", "102"). Code activities. First criterion for identifying activities that must have fishing time

objectoperation_fishingtime

character expected. Default values: c(NA, NA, NA, "99", NA, NA, NA, "1", "2", "4", "6", "8", "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "99", NA). Code object operation. Second criterion for identifying activities that must have fishing time (Indicate NA if you want the code object operation to be missing)

Value

The function returns a character with output is "message", a data.frame with output is "report", a logical with output is "logical"

Details

The input dataframe must contain all these columns for the function to work :

route_id

route_seatime

route_fishingtime Dataframe 2:

activity_id

vesselactivity_code

route_id Dataframe 3:

floatingobject_id

objectoperation_code

activity_id