Function to create a data.frame in list

Function to create a data.frame in list

data_to_list(
  data,
  name_col_dataplot,
  colname_id,
  colname_plot,
  colname_info,
  rename_colname_info
)

Arguments

data

data.frame expected. Csv or output of the function data_extraction, which must be done before using the data_to_list () function.

name_col_dataplot

character expected. Name of the sub-list containing the table

colname_id

character expected. Name of the column in data used to identify a group

colname_plot

character expected. Vector of column names in data to be grouped in an array because it contains several values per group

colname_info

character expected. Vector of column names in data to be organized in separate sub-lists because it contains a single value per group

rename_colname_info

character expected. Vector with sub-list names to be renamed

Value

The function returns a list with the sub list

Examples

dataframe1 <- data.frame(trip_id = c("1", "1", "2", "2", "2"),
                         trip_startdate = as.Date(c("2020/01/01", "2020/01/01", "2020/01/01",
                                                    "2020/01/01", "2020/01/01")),
                         trip_enddate = as.Date(c("2020/01/02", "2020/01/02", "2020/01/02",
                                                  "2020/01/02", "2020/01/02")),
                         activity_date = as.Date(c("2020/01/01", "2020/01/02","2020/01/01",
                                                   "2020/01/02", "2020/01/03")),
                         logical = c(TRUE, TRUE, TRUE, TRUE, FALSE))
data_to_list(data = dataframe1, name_col_dataplot = "data", colname_id = "trip_id",
             colname_plot = c("activity_date", "logical"),
             colname_info = c("trip_startdate", "trip_enddate"),
             rename_colname_info = c("startdate", "enddate"))
#> $`1`
#> $`1`$data
#>   activity_date logical
#> 1    2020-01-01    TRUE
#> 2    2020-01-02    TRUE
#> 
#> $`1`$startdate
#> [1] "2020-01-01"
#> 
#> $`1`$enddate
#> [1] "2020-01-02"
#> 
#> 
#> $`2`
#> $`2`$data
#>   activity_date logical
#> 3    2020-01-01    TRUE
#> 4    2020-01-02    TRUE
#> 5    2020-01-03   FALSE
#> 
#> $`2`$startdate
#> [1] "2020-01-01"
#> 
#> $`2`$enddate
#> [1] "2020-01-02"
#> 
#>