tab_content Server Functions

A shiny Module for creation of all tab

mod_tab_content_server(
  id,
  tab_info,
  check_info,
  type_check_info,
  config_data,
  res_auth
)

Arguments

id

Internal parameters for shiny

tab_info

list expected. Reactive value contains information about the dynamic tab display allowed for user type

check_info

list expected. Reactive value contains information about the dynamic check display allowed for user type

type_check_info

list expected. Information about the dynamic type of check

config_data

list expected. Reactive value contains the contents of the configuration file

res_auth

character expected. Reactive value containing information about the user's connection

Value

The function returns nothing, instantiating the tab

Details

tab_info_authorize :

id : Tab identification, mandatory, character expected

check_info:

id_check_authorize : Check identification, mandatory, character expected

check_info_authorize :

id : Check identification, mandatory, character expected

tab : Tab identification, mandatory, character expected

title : Name to be displayed for check, optional, character expected

text : Text to be displayed for check, optional, character expected

type : Check type, you must choose between check type identifiers (type_check_info), mandatory, character expected

column_no_wrap : Column numbers that should not be subject to automatic line breaks, optional, integer expected

size_box : Check size specification for each window size type (format "col-sm-your_size col-md-your_size col-lg-your_size"), optional, character expected

id : Button selection identification, mandatory, character expected

specific_check : Indicates whether the selection should display specific controls or not, optional (default : TRUE), logical expected

databases_configuration : named list with database connection information