table Server Functions

A shiny Module for creation of all table

mod_table_server(
  id,
  data,
  name,
  type_line_check,
  referential_file,
  column_no_wrap = NULL,
  name_function_plot = NULL,
  name_function_text_plot = NULL,
  title_window = NULL
)

Arguments

id

Internal parameters for shiny

data

list expected. Reactive list containing results tables for all checks

name

character expected. Control identifier in data to be displayed

type_line_check

character expected. Reactive value containing the type of line selected by the user

referential_file

list expected. Reactive list containing referential tables for all plot

column_no_wrap

integer expected. Default values: NULL. Column numbers that should not be subject to automatic line breaks

name_function_plot

character expected. Default values: NULL. Name of the function that creates the plot

name_function_text_plot

character expected. Default values: NULL. Name of the function that creates the text to be displayed in the plot window

title_window

character expected. Default values: NULL. Plot window name

Value

The function returns nothing, instantiating the table