Function to create the plot of the consistency of the position for the activity and VMS

Function to create the plot

plot_anapo(data_vms, crs_vms, crs_activity, data_activity, data_trip)

Arguments

data_vms

data.frame expected. Csv or an element of the output of function plot_anapo_data, which must be done before using the plot_anapo () function.

crs_vms

numeric expected. Coordinates Reference System for position VMS

crs_activity

numeric expected. Coordinates Reference System for position activity

data_activity

data.frame expected. Csv or an element of the output of function plot_anapo_data, which must be done before using the plot_anapo () function.

data_trip

data.frame expected. Csv or an element of the output of function plot_anapo_data, which must be done before using the plot_anapo () function.

Value

The function returns a plotly # Example 1 data_vms <- data.frame(vms_date = as.Date(c("2020/01/12", "2020/01/13", "2020/01/13")), vms_time = c("15:26:01", "10:55:15", "22:32:17"), vms_position = c("POINT (4 4)", "POINT (0 0.1)", "POINT (3 0.3)"), distance = c(339.502583, 6.004055, 181.019203), duration = c(NA, NA, NA), score = c(NA_real_, NA_real_, NA_real_)) units::install_unit("NM", "1852 m", "Nautical mile") data_vms$distance <- units::set_units(data_vms$distance, NM) units::install_unit("ms", "1000 secs", "Milliseconds") data_vms$duration <- units::set_units(data_vms$duration, ms) data_activity <- data.frame(activity_date = as.Date(c("2020/01/12")), activity_number = 1L, activity_position = "POINT (0 0)", vesselactivity_code = "2", grounding = FALSE) data_trip <- data.frame(activity_date = as.Date(c("2020/01/12", "2020/01/13", "2020/01/13")), activity_time = c("05:26:01", "03:12:34", "23:26:47"), activity_number = c(1L, 1L, 1L), activity_position = c("POINT (0 0)", "POINT (2.6 3.8)", "POINT (3 0.6)"), vesselactivity_code = c("2", "6", "13"), grounding = c(FALSE,FALSE, TRUE)) plot_anapo(data_vms, 4326, 4326, data_activity, data_trip)

Details

The input dataframe must contain all these columns for the function to work :

vms_date

vms_time

vms_position

distance

duration

score

activity_date

activity_number

activity_position

vesselactivity_code

grounding

activity_date

activity_time

activity_number

activity_position

vesselactivity_code

grounding