Function to create list data/argument for the plot plot_anapo

Function to create list data/argument

plot_anapo_data(dataframe1, dataframe2, dataframe3)

Arguments

dataframe1

data.frame expected. Csv or the second output of the function logbook_anapo_control, which must be done before using the plot_anapo_data () function.

dataframe2

data.frame expected. Csv or output of the function data_extraction, which must be done before using the plot_anapo_data () function.

dataframe3

data.frame expected. Csv or output of the function data_extraction, which must be done before using the plot_anapo_data () function.

Value

The function returns a list with the sub list is a data for one plot

Details

The input dataframe must contain all these columns for the function to work :

activity_id

activity_date

activity_time

activity_position

activity_crs

vms_date

vms_time

vms_position

distance

duration

score

vms_crs

trip_id

vessel_code

trip_enddate

activity_id

activity_date

activity_number

vesselactivity_code

transmittingbuoy_id

transmittingbuoyoperation_code

activity_id

Examples

dataframe1 <- data.frame(activity_id = c("4", "4", "4", "6", "6", "6", "2", "2", "2"),
                         activity_date = as.Date(c("2020/01/13", "2020/01/13", "2020/01/13",
                                                   "2020/01/13", "2020/01/13", "2020/01/13",
                                                   "2020/01/12", "2020/01/12", "2020/01/12")),
                         activity_time = c("03:12:34", "03:12:34", "03:12:34", "23:26:47",
                                           "23:26:47", "23:26:47", "05:26:01", "05:26:01",
                                           "05:26:01"),
                         activity_position = c("POINT (2.6 3.8)", "POINT (2.6 3.8)",
                                               "POINT (2.6 3.8)", "POINT (3 0.6)",
                                               "POINT (3 0.6)", "POINT (3 0.6)",
                                               "POINT (0 0)", "POINT (0 0)", "POINT (0 0)"),
                         activity_crs = c(4326, 4326, 4326, 4326, 4326, 4326, 4326, 4326, 4326),
                         vms_date = as.Date(c("2020/01/13", "2020/01/13", "2020/01/12",
                                              "2020/01/13", "2020/01/13", "2020/01/12",
                                              "2020/01/12", "2020/01/13", "2020/01/13")),
                         vms_time = c("10:55:15", "22:32:17", "15:26:01", "10:55:15",
                                      "22:32:17", "15:26:01", "15:26:01", "10:55:15", "22:32:17"),
                         vms_position = c("POINT (0 0.1)", "POINT (3 0.3)", "POINT (4 4)",
                                          "POINT (0 0.1)", "POINT (3 0.3)", "POINT (4 4)",
                                          "POINT (4 4)", "POINT (0 0.1)", "POINT (3 0.3)"),
                         distance = c(271.445826973379, 211.507663354955, 84.71741855987,
                                      182.602328607533, 18.012165417519, 212.768118823317,
                                      339.502583, 6.004055, 181.019203),
                         duration = c(-27761000, -69583000, -44007000, 45092000, 3270000,
                                      28846000, NA, NA, NA),
                         score = c(0, 0, 0, 0, 0.209441144555651, 0, NA, NA, NA),
                         vms_crs = c(4326, 4326, 4326, 4326, 4326, 4326, 4326, 4326, 4326))
units::install_unit("NM", "1852 m", "Nautical mile")
dataframe1$distance <- units::set_units(dataframe1$distance, NM)
units::install_unit("ms", "1000 secs", "Milliseconds")
dataframe1$duration <- units::set_units(dataframe1$duration, ms)
dataframe2 <- data.frame(trip_id  = c("1", "1", "1"),
                         vessel_code = c("0", "0", "0"),
                         trip_enddate = as.Date(c("2020/01/20", "2020/01/20", "2020/01/20")),
                         activity_id = c("2", "4", "6"),
                         activity_date = as.Date(c("2020/01/12", "2020/01/13", "2020/01/13")),
                         activity_number = c(1L, 1L, 1L),
                         vesselactivity_code = c("2","6", "13"))
dataframe3 <- data.frame(transmittingbuoy_id  = c("1"),
                         transmittingbuoyoperation_code = c("4"),
                         activity_id = c("6"))
plot_anapo_data(dataframe1, dataframe2, dataframe3)
#> $`2`
#> $`2`$data_trip
#> # A tibble: 3 × 6
#>   activity_date activity_time activity_position activity_number grounding
#>   <date>        <chr>         <chr>                       <int> <lgl>    
#> 1 2020-01-12    05:26:01      POINT (0 0)                     1 FALSE    
#> 2 2020-01-13    03:12:34      POINT (2.6 3.8)                 1 FALSE    
#> 3 2020-01-13    23:26:47      POINT (3 0.6)                   1 TRUE     
#> # ℹ 1 more variable: vesselactivity_code <chr>
#> 
#> $`2`$data_activity
#>   vessel_code trip_enddate activity_date activity_time activity_position
#> 3           0   2020-01-20    2020-01-12      05:26:01       POINT (0 0)
#>   activity_number grounding vesselactivity_code     X     Y
#> 3               1     FALSE                   2 E0°0' N0°0'
#> 
#> $`2`$data_vms
#>    vms_position   vms_date vms_time        distance duration score
#> 7   POINT (4 4) 2020-01-12 15:26:01 339.502583 [NM]  NA [ms]    NA
#> 8 POINT (0 0.1) 2020-01-13 10:55:15   6.004055 [NM]  NA [ms]    NA
#> 9 POINT (3 0.3) 2020-01-13 22:32:17 181.019203 [NM]  NA [ms]    NA
#> 
#> $`2`$crs_activity
#> [1] 4326
#> 
#> $`2`$crs_vms
#> [1] 4326
#> 
#> 
#> $`4`
#> $`4`$data_trip
#> # A tibble: 3 × 6
#>   activity_date activity_time activity_position activity_number grounding
#>   <date>        <chr>         <chr>                       <int> <lgl>    
#> 1 2020-01-13    03:12:34      POINT (2.6 3.8)                 1 FALSE    
#> 2 2020-01-13    23:26:47      POINT (3 0.6)                   1 TRUE     
#> 3 2020-01-12    05:26:01      POINT (0 0)                     1 FALSE    
#> # ℹ 1 more variable: vesselactivity_code <chr>
#> 
#> $`4`$data_activity
#>   vessel_code trip_enddate activity_date activity_time activity_position
#> 1           0   2020-01-20    2020-01-13      03:12:34   POINT (2.6 3.8)
#>   activity_number grounding vesselactivity_code      X      Y
#> 1               1     FALSE                   6 E2°36' N3°48'
#> 
#> $`4`$data_vms
#>    vms_position   vms_date vms_time       distance       duration score
#> 1 POINT (0 0.1) 2020-01-13 10:55:15 271.44583 [NM] -27761000 [ms]     0
#> 2 POINT (3 0.3) 2020-01-13 22:32:17 211.50766 [NM] -69583000 [ms]     0
#> 3   POINT (4 4) 2020-01-12 15:26:01  84.71742 [NM] -44007000 [ms]     0
#> 
#> $`4`$crs_activity
#> [1] 4326
#> 
#> $`4`$crs_vms
#> [1] 4326
#> 
#> 
#> $`6`
#> $`6`$data_trip
#> # A tibble: 3 × 6
#>   activity_date activity_time activity_position activity_number grounding
#>   <date>        <chr>         <chr>                       <int> <lgl>    
#> 1 2020-01-13    03:12:34      POINT (2.6 3.8)                 1 FALSE    
#> 2 2020-01-13    23:26:47      POINT (3 0.6)                   1 TRUE     
#> 3 2020-01-12    05:26:01      POINT (0 0)                     1 FALSE    
#> # ℹ 1 more variable: vesselactivity_code <chr>
#> 
#> $`6`$data_activity
#>   vessel_code trip_enddate activity_date activity_time activity_position
#> 2           0   2020-01-20    2020-01-13      23:26:47     POINT (3 0.6)
#>   activity_number grounding vesselactivity_code     X      Y
#> 2               1      TRUE                  13 E3°0' N0°36'
#> 
#> $`6`$data_vms
#>    vms_position   vms_date vms_time       distance      duration     score
#> 4 POINT (0 0.1) 2020-01-13 10:55:15 182.60233 [NM] 45092000 [ms] 0.0000000
#> 5 POINT (3 0.3) 2020-01-13 22:32:17  18.01217 [NM]  3270000 [ms] 0.2094411
#> 6   POINT (4 4) 2020-01-12 15:26:01 212.76812 [NM] 28846000 [ms] 0.0000000
#> 
#> $`6`$crs_activity
#> [1] 4326
#> 
#> $`6`$crs_vms
#> [1] 4326
#> 
#>