Function to create the plot of the consistency of the eez for the activity

Function to create the plot

plot_eez(data, referential_geographical_shape)

Arguments

data

data.frame expected. Csv or an element of the output of function plot_eez_data, which must be done before using the plot_eez () function.

referential_geographical_shape

data.frame expected. Layer to containing the eez shapefile (example cf : Flanders Marine Institute (2023). Maritime Boundaries Geodatabase: Maritime Boundaries and Exclusive Economic Zones (200NM), version 12. Available online at https://www.marineregions.org/. https://doi.org/10.14284/632)

Value

The function returns a plotly # Example 1 data <- data.frame(activity_position = "POINT (0.1 0)", activity_crs = 4326) referential_geographical_shape <- sf::st_sf(data.frame(ISO_TER1 = c("SYC", "XSG"), ISO_TER2 = c(NA, NA), ISO_TER3 = c(NA, NA), geometry = sf::st_sfc(sf::st_polygon(list(rbind(c(0,0), c(2,0), c(2,2), c(0,2), c(0,0)))), sf::st_polygon(list(rbind(c(3,3), c(3,5), c(5,5), c(5,3), c(3,3)))), crs = 4326))) plot_eez(data, referential_geographical_shape)

Details

The input dataframe must contain all these columns for the function to work :

activity_position

activity_crs

ISO_TER1

ISO_TER2

ISO_TER3

geometry