Function to create list data/argument for the plot eez

Function to create list data/argument

plot_eez_data(dataframe1, dataframe2)

Arguments

dataframe1

data.frame expected. Csv or the second output of the function check_eez_inspector, which must be done before using the plot_eez_data () function.

dataframe2

data.frame expected. Csv or output of the function data_extraction, which must be done before using the plot_eez_data () function.

Value

The function returns a list with the sub list is a data for one plot

Details

The input dataframe must contain all these columns for the function to work :

activity_id

activity_position

activity_crs

fpazone_code

fpazone_country_iso3

eez_calculated

activity_id

vessel_code

trip_enddate

activity_date

activity_number

Examples

dataframe1 <- data.frame(activity_id = c("1", "2"),
                         activity_position = c("POINT (0.1 0)", "POINT (0 0)"),
                         fpazone_code = c("SYC", "XSG"),
                         fpazone_country_iso3 = c("SYC", "XXX"),
                         eez_calculated = c("SYC", "XSG"),
                         activity_crs = c(4326, 4326))
dataframe2 <- data.frame(activity_id = c("1", "2"),
                         vessel_code = c("0", "0"),
                         trip_enddate = as.Date(c("2020/01/10", "2020/01/10")),
                         activity_date = as.Date(c("2020/01/09", "2020/01/10")),
                         activity_number = c(1L, 1L))
plot_eez_data(dataframe1, dataframe2)
#> $`1`
#> $`1`$data
#>   activity_position activity_crs
#> 1     POINT (0.1 0)         4326
#> 
#> $`1`$vessel_code
#> [1] "0"
#> 
#> $`1`$trip_enddate
#> [1] "2020-01-10"
#> 
#> $`1`$activity_date
#> [1] "2020-01-09"
#> 
#> $`1`$activity_number
#> [1] 1
#> 
#> $`1`$fpazone_code
#> [1] "SYC"
#> 
#> $`1`$fpazone_country_iso3
#> [1] "SYC"
#> 
#> $`1`$eez_calculated
#> [1] "SYC"
#> 
#> $`1`$X
#> [1] "E0°6'"
#> 
#> $`1`$Y
#> [1] "N0°0'"
#> 
#> $`1`$referential_geographical_shape
#> [1] "shape_eez"
#> 
#> 
#> $`2`
#> $`2`$data
#>   activity_position activity_crs
#> 2       POINT (0 0)         4326
#> 
#> $`2`$vessel_code
#> [1] "0"
#> 
#> $`2`$trip_enddate
#> [1] "2020-01-10"
#> 
#> $`2`$activity_date
#> [1] "2020-01-10"
#> 
#> $`2`$activity_number
#> [1] 1
#> 
#> $`2`$fpazone_code
#> [1] "XSG"
#> 
#> $`2`$fpazone_country_iso3
#> [1] "XXX"
#> 
#> $`2`$eez_calculated
#> [1] "XSG"
#> 
#> $`2`$X
#> [1] "E0°0'"
#> 
#> $`2`$Y
#> [1] "N0°0'"
#> 
#> $`2`$referential_geographical_shape
#> [1] "shape_eez"
#> 
#>