Function to create the text for windows with plot eez

Function to create the windows with plot of the consistency of the eez for the activity

plot_eez_windows(
  vessel_code,
  trip_enddate,
  activity_date,
  activity_number,
  X,
  Y,
  fpazone_code,
  fpazone_country_iso3,
  eez_calculated
)

Arguments

vessel_code

character expected. Vessel code

trip_enddate

Date expected. Trip end date

activity_date

Date expected. Activity date

activity_number

integer expected. Activity number

X

character expected. Position longitude

Y

character expected. Position latitude

fpazone_code

character expected. Fpa zone code

fpazone_country_iso3

character expected. Iso3 country code for fpa zone

eez_calculated

character expected. Calculated zee zone code

Value

The function returns a character

Examples

plot_eez_windows(vessel_code = "0", trip_enddate = as.Date("2020/01/10"),
                 activity_date = as.Date("2020/01/09"), activity_number = 1L, X = "E0°6'",
                 Y = "N0°0'", fpazone_code = "SYC", fpazone_country_iso3 = "SYC",
                 eez_calculated = "SYC")
#> [1] "<b>Trip information : </b><br>\n         <ul><li>Vessel code : 0</li>\n         <li>Trip end date : 2020-01-10</li>\n         <li>Activity date : 2020-01-09</li>\n         <li>Activity number : 1</li>\n         <li>Latitude : N0°0'</li>\n         <li>Longitude : E0°6'</li></ul>\n         <b>Problem information : </b><br>\n         <ul><li>Declared eez : SYC</li>\n         <li>Declared country eez : SYC</li>\n         <li>Calculated eez : SYC</li></ul>"